+++ horovod/common/timeline.cc
@@ -24,26 +24,99 @@
 namespace horovod {
 namespace common {
 
-void TimelineWriter::Initialize(std::string file_name) {
-  file_.open(file_name, std::ios::out | std::ios::trunc);
+TimelineWriter::TimelineWriter(){
+  cur_filename_ = "";
+  new_pending_filename_ = "";
+}
+
+void TimelineWriter::SetPendingTimelineFile(std::string filename){
+    std::lock_guard<std::recursive_mutex> guard(writer_mutex_);
+    new_pending_filename_ = filename;
+}
+
+std::string TimelineWriter::PendingTimelineFile(){
+    std::lock_guard<std::recursive_mutex> guard(writer_mutex_);
+    return new_pending_filename_;
+}
+
+void TimelineWriter::SetTimelineFile(std::string filename){
+  LOG(INFO) << "Setting TimelineFile. Current file:" << cur_filename_ << " New filename:" << filename;
+  // Close if there existing file open and new file is not same as existing file
+  if (cur_filename_ != "" && cur_filename_ != filename){
+      if (file_.is_open() && file_.good()){
+        file_.close();
+        LOG(INFO) << "Closed timeline file:" << cur_filename_;
+      }
+      tensor_table_.clear();
+  }
+  // if new filename is empty, we need to stop accepting activities. This would stopping timeline
+  if(filename == "") {
+    cur_filename_ = filename;
+    new_pending_filename_ = cur_filename_;
+    // empty filename is special which tells that init the timeline but don't activate it.
+    healthy_ = true;
+    active_ = false;
+    LOG(INFO) << "Inited TimelineWriter but active_ is false, since filename passed is empty string";
+    return;
+  }
+  // if newfile is same as existing file, it should be no-op and we can continue writing in existing file.
+  if(cur_filename_ == filename) {
+    if (file_.good()) {
+      LOG(INFO) << "Timeline cur_filename_ is same as existing filename:"
+                << filename << " Set active and healthy to true" ;
+      cur_filename_ = filename;
+      new_pending_filename_ = cur_filename_;
+      is_new_file_ = false;
+      healthy_ = true;
+      active_ = true;
+      return;
+    } else {
+        LOG(ERROR) << "Filename:" << filename
+                  << " is not good. Setting healthy to true and active to false since file is not good";
+        cur_filename_ = filename;
+        new_pending_filename_ = cur_filename_;
+        healthy_ = true;
+        active_ = false;
+    }
+  }
+  // all other cases, need to create a new file
+  file_.open(filename, std::ios::out | std::ios::trunc);
   if (file_.good()) {
-    // Initialize the timeline with '[' character.
-    file_ << "[\n";
+    LOG(INFO) << "Opened new timeline file" << filename << " Set active and healthy to true";
+    cur_filename_ = filename;
+    new_pending_filename_ = cur_filename_;
+    is_new_file_ = true;
     healthy_ = true;
     active_ = true;
-
-    // Spawn writer thread.
-    writer_thread_ = std::thread(&TimelineWriter::WriterLoop, this);
   } else {
-    LOG(ERROR) << "Error opening the Horovod Timeline file " << file_name
+    LOG(ERROR) << "Error opening the Horovod Timeline file " << filename
                << ", will not write a timeline.";
+    healthy_ = true;
+    active_ = false;
   }
+}
+
+void TimelineWriter::Initialize(std::string file_name) {
+  SetTimelineFile(file_name);
+  // Spawn writer thread.
+  writer_thread_ = std::thread(&TimelineWriter::WriterLoop, this);
 }
 
 void TimelineWriter::Shutdown() {
   active_ = false;
-  writer_thread_.join();
-  file_.close();
+  healthy_ = false;
+  try {
+    if (writer_thread_.joinable()){
+        writer_thread_.join();
+    }
+  } catch(const std::system_error& e) {
+    LOG(INFO) << "Caught system_error while joining writer thread. Code " << e.code()
+              << " meaning " << e.what();
+  }
+
+  if (cur_filename_ != "" && file_.is_open() && file_.good()){
+    file_.close();
+  }
   tensor_table_.clear();
 }
 
@@ -76,7 +149,15 @@ void TimelineWriter::EnqueueWriteMarker(const std::string& name,
 
 void TimelineWriter::DoWriteEvent(const TimelineRecord& r) {
   assert(r.type == TimelineRecordType::EVENT);
-
+  if(is_new_file_){
+    file_ << "[\n";
+    is_new_file_ = false;
+  } else {
+    // last event closed the json ']' , need to seek to one position back and write ',' to continue
+    long pos = file_.tellp();
+    file_.seekp (pos-1);
+    file_ << ",";
+  }
   auto& tensor_idx = tensor_table_[r.tensor_name];
   if (tensor_idx == 0) {
     tensor_idx = (int)tensor_table_.size();
@@ -110,23 +191,34 @@ void TimelineWriter::DoWriteEvent(const TimelineRecord& r) {
   if (r.args != "") {
     file_ << ", \"args\": {" << r.args << "}";
   }
-  file_ << "}," << std::endl;
+  // We make sure that the events are written always produce valid json file
+  file_ << "}]";
 }
 
 void TimelineWriter::DoWriteMarker(const TimelineRecord& r) {
   assert(r.type == TimelineRecordType::MARKER);
-
+  if(is_new_file_){
+    file_ << "[\n";
+    is_new_file_ = false;
+  } else {
+    // last event closed the json ']' , need to seek to one position back and write ',' to continue
+    long pos = file_.tellp();
+    file_.seekp (pos-1);
+    file_ << ",";
+  }
   file_ << "{";
   file_ << "\"ph\": \"i\"";
   file_ << ", \"name\": \"" << r.marker_name << "\"";
   file_ << ", \"ts\": " << r.ts_micros << "";
   file_ << ", \"s\": \"g\"";
-  file_ << "}," << std::endl;
+  // We make sure that the events are written always produce valid json file
+  file_ << "}]";
+
 }
 
 void TimelineWriter::WriterLoop() {
-  while (healthy_ && active_) {
-    while (healthy_ && !record_queue_.empty()) {
+  while (healthy_) {
+    while (healthy_ && active_ && !record_queue_.empty()) {
       auto& r = record_queue_.front();
       switch (r.type) {
       case TimelineRecordType::EVENT:
@@ -140,13 +232,23 @@ void TimelineWriter::WriterLoop() {
       }
       record_queue_.pop();
 
-      if (!file_.good()) {
+      if (active_ && !file_.good()) {
         LOG(ERROR) << "Error writing to the Horovod Timeline after it was "
                       "successfully opened, will stop writing the timeline.";
-        healthy_ = false;
+        active_ = false;
       }
     }
-
+    if(!active_){
+      // drain record_queue
+      while(!record_queue_.empty()){
+        record_queue_.pop();
+      }
+    }
+    // check if we need to call SetTimeLineFile
+    std::string pending_timeline_file = PendingTimelineFile();
+    if (pending_timeline_file != cur_filename_) {
+        SetTimelineFile(pending_timeline_file);
+    }
     // Allow scheduler to schedule other work for this core.
     std::this_thread::yield();
   }
@@ -156,9 +258,8 @@ void Timeline::Initialize(std::string file_name, unsigned int horovod_size) {
   if (initialized_) {
     return;
   }
-
   // Start the writer.
-  writer_.Initialize(std::move(file_name));
+  writer_.Initialize(file_name);
 
   // Initialize if we were able to open the file successfully.
   initialized_ = writer_.IsHealthy();
@@ -320,5 +421,9 @@ void Timeline::MarkCycleStart() {
   WriteMarker("CYCLE_START");
 }
 
+void Timeline::SetPendingTimelineFile(std::string filename) {
+    writer_.SetPendingTimelineFile(filename);
+    LOG(INFO) << "Set pending timeline file to " << filename;
+}
 } // namespace common
 } // namespace horovod